+++ test/integration/voteRouting.test.js
@@ -9,9 +9,11 @@ chai.use(chaiHttp)
 chai.use(dirtyChai)
 const server = require('../../server/server')
 
-// // require the database models
+// require the database models
 const models = require('../../server/db/models')
-const query = require('../../server/queryAPI/apiQuery')
+const checkEmptyDatabase = require('../helper').checkEmptyDatabase
+const userQuery = require('../../server/queryAPI').userQuery
+const quizQuery = require('../../server/queryAPI').quizQuery
 
 const title =
 `
@@ -47,27 +49,15 @@ describe(title, () => {
   before(() => {
     return models.sequelize.sync({ force: true })
     .then(() => {
-    // 1. Make FIND ALL queries into all tables
-      return Promise.all([
-        query.findAllUsers(),
-        query.findAllQuizzes(),
-        query.findAllVotes(),
-        query.findAllQuizzesTaken()
-      ])
-    })
-    .then((promiseArray) => {
-    // 2. check all queries to see if they are empty []
-      promiseArray.forEach((searchResult) => {
-        assert.deepEqual(searchResult, [])
-      })
+      checkEmptyDatabase()
     })
     .then(() => {
     // 3. add data into other tables HERE
-      return query.newUser(user1)
+      return userQuery.createUser(user1)
     }).then((newUserPromise) => {
       assert.isTrue(newUserPromise[1], 'user should be created')
     }).then(() => {
-      return query.newQuiz(quiz1)
+      return quizQuery.createQuiz(quiz1)
     }).then((insertPromise) => {
       assert.isTrue(insertPromise[0][1], 'quiz should be created')
     })