+++ test/unit/question.test.js
@@ -4,7 +4,10 @@ const assert = require('chai').assert
 const expect = require('chai').expect
 
 const models = require('../../server/db/models')
-const query = require('../../server/queryAPI/apiQuery')
+const checkEmptyDatabase = require('../helper').checkEmptyDatabase
+const userQuery = require('../../server/queryAPI').userQuery
+const quizQuery = require('../../server/queryAPI').quizQuery
+const questionQuery = require('../../server/queryAPI').questionQuery
 
 const title =
 `
@@ -42,41 +45,26 @@ describe(title, () => {
   before(() => {
     return models.sequelize.sync({ force: true })
     .then(() => {
-      // 1. Make FIND ALL queries into all tables
-      return Promise.all([
-        query.findAllUsers(),
-        query.findAllQuizzes(),
-        query.findAllVotes(),
-        query.findAllQuizzesTaken()
-      ])
-    })
-    .then((promiseArray) => {
-      // 2. check all queries to see if they are empty []
-      promiseArray.forEach((searchResult) => {
-        assert.deepEqual(searchResult, [])
-      })
+      checkEmptyDatabase()  
     })
     .then(() => {
       // 3. add data into other tables HERE
-      return Promise.all([
-        query.newUser(user1)
-      ])
+      return userQuery.createUser(user1)
     })
-    .then((insertPromises) => {
+    .then((queryPromise) => {
       // 4. check that all insertions worked!
-      assert.isTrue(insertPromises[0][1], 'user should be created')
-      // assert.isTrue(insertPromises[1][1], 'quiz should be created')
-      return query.newQuiz(quiz1)
+      assert.isTrue(queryPromise[1], 'user should be created')
+      return quizQuery.createQuiz(quiz1)
     })
-    .then((insertPromise) => {
-      assert.isTrue(insertPromise[0][1], 'quiz should be created')
+    .then((newQuiz) => {
+      assert.isTrue(newQuiz[1], 'quiz should be created')
     })
   })
     /** ================== Actual Tests Begin Here ========================
    *
    */
   it('should be able add a question to quiz 1', (done) => {
-    query.addQuestion(question1).then((r) => {
+    questionQuery.createQuestion(question1).then((r) => {
       const result = JSON.parse(JSON.stringify(r))
       // console.log(result)
       // console.log(typeof result.choices) // object!
@@ -88,7 +76,7 @@ describe(title, () => {
   })
 
   it('should be able to add another question to quiz 1', (done) => {
-    query.addQuestion(question2).then((r) => {
+    questionQuery.createQuestion(question2).then((r) => {
       const result = JSON.parse(JSON.stringify(r))
       // console.log(result)
       expect(result).to.include.keys([
@@ -100,7 +88,7 @@ describe(title, () => {
 
   it('should be able to find both questions of quiz 1', (done) => {
     // done()
-    query.findQuizById(1).then((rawResult) => {
+    quizQuery.findQuizById(1).then((rawResult) => {
       const result = JSON.parse(JSON.stringify(rawResult))
       // console.log(result)
       expect(result.questions).to.have.lengthOf(2)