+++ test/unit/user.test.js
@@ -4,7 +4,8 @@ const assert = require('chai').assert
 const expect = require('chai').expect
 
 const models = require('../../server/db/models')
-const query = require('../../server/queryAPI/apiQuery')
+const checkEmptyDatabase = require('../helper').checkEmptyDatabase
+const userQuery = require('../../server/queryAPI').userQuery
 
 const title =
 `
@@ -20,50 +21,22 @@ let user1 = {
 }
 
 describe(title, () => {
-  // before(function () {
-  //   return new Promise((resolve, reject) => {
-  //     models.sequelize.query('SET FOREIGN_KEY_CHECKS = 0', {raw: true}).then(() => {
-  //       models.sequelize.sync({ force: true }).then(() => {
-  //         resolve()
-  //       })
-  //     })
-  //   })
-  // })
   before(() => {
-    // return models.sequelize.query('SET FOREIGN_KEY_CHECKS = 0', {raw: true})
-    // .then(() => {
-    //   return models.sequelize.sync({ force: true })
-    // })
     return models.sequelize.sync({ force: true })
     .then(() => {
-    // 1. Make FIND ALL queries into all tables
-      return Promise.all([
-        query.findAllUsers(),
-        query.findAllQuizzes(),
-        query.findAllVotes(),
-        query.findAllQuizzesTaken()
-      ])
-    })
-    .then((promiseArray) => {
-    // 2. check all queries to see if they are empty []
-      promiseArray.forEach((searchResult) => {
-        assert.deepEqual(searchResult, [])
-      })
-    })
-    .then(() => {
-    // 3. add data into other tables HERE
+      checkEmptyDatabase()
     })
   })
 
   it('Should not return a non-existent user', (done) => {
-    query.findUserById(-99).then((results) => {
+    userQuery.findUserById(-99).then((results) => {
       expect(results).to.be.a('null')
       done()
     })
   })
 
   it('should be able to create a new user', (done) => {
-    query.newUser(user1).spread((result, created) => {
+    userQuery.createUser(user1).spread((result, created) => {
       try {
         assert.isTrue(created, 'user was created')
         done()
@@ -74,7 +47,7 @@ describe(title, () => {
   })
 
   it('should not be able to create the same user twice', (done) => {
-    query.newUser(user1).spread((result, created) => {
+    userQuery.createUser(user1).spread((result, created) => {
       try {
         assert.isNotTrue(created, 'duplicate use should not have been created again')
         done()
@@ -85,7 +58,7 @@ describe(title, () => {
   })
 
   it('should be able to find the user that was created', (done) => {
-    query.findUserById(1).then((result) => {
+    userQuery.findUserById(1).then((result) => {
       try {
         let user = JSON.parse(JSON.stringify(result))
         // let user = result.dataValues
@@ -100,7 +73,7 @@ describe(title, () => {
   })
 
   it('should be able to check the hash password correctly', (done) => {
-    query.findUserByUsername(user1.username)
+    userQuery.findUserByUsername(user1.username)
     .then((userResult) => {
       const incorrectPass = userResult.comparePassword('wrong password')
       const correctPass = userResult.comparePassword(user1.password)